move bmi3d functions to separate module
